add workspace member permission

+++ Ticket.API/Controllers/WorkSpaceMembersController.cs
@@ -5,10 +5,12 @@
     [Route("api/[controller]")]
     public class WorkSpaceMembersController : BaseController
     {
+        private readonly IAuthorizationService _authService;
         private readonly IWorkSpaceMemberService _workSpaceMemberService;
 
-        public WorkSpaceMembersController(IWorkSpaceMemberService workSpaceMemberService)
+        public WorkSpaceMembersController(IAuthorizationService authService, IWorkSpaceMemberService workSpaceMemberService)
         {
+            _authService = authService;
             _workSpaceMemberService = workSpaceMemberService;
         }
 
@@ -22,6 +24,10 @@
             [FromRoute] string workSpaceId,
             [FromQuery] WorkSpaceMemberRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.GetListWorkSpaceMember);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             var res = await _workSpaceMemberService.GetWorkSpaceMembers(model, workSpaceId);
             return SuccessWithPagination(res.Pagination, res.Members);
         }
@@ -34,6 +40,10 @@
         [HttpPost]
         public async Task<BaseResponse> AddNewMemberInWorkSpace([FromBody] WorkSpaceAddMemberRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.AddWorkSpaceMember);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             await _workSpaceMemberService.AddMemberInWorkSpace(model, User.Identity.Name);
             return Success();
         }
@@ -46,6 +56,10 @@
         [HttpDelete]
         public async Task<BaseResponse> DeleteExistMemberInWorkSpace([FromBody] WorkSpaceRemoveMemberRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.RemoveWorkSpaceMember);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             await _workSpaceMemberService.RemoveMemberInWorkSpace(model, User.Identity.Name);
             return Success();
         }
